do not popup reboot dialog while in xbian-config xbmc
template for "skin.amber" (not fully working)

+++ content/home/xbian/.xbmc/addons/plugin.xbianconfig/services/upgrade.py
@@ -49,6 +49,9 @@ class upgrade(service):
         if self.rebootNoCheck:
             return
 
+        if os.path.isfile('/tmp/.xbian_config_python'):
+            return 
+
         rebootneeded = xbianConfig('reboot')
 
         if rebootneeded and rebootneeded[0] == '1' :
